Start packed-Boolean collection and support types
Add a collection of Bool that stores its elements as bits within a
collection of unsigned integers.  Start with the stored properties and
their types and a master initializer.

Add an enumeration type so the packed-Boolean collection can flag which
way bits should be read within words during an initializer.

Extend some of the Integer support protocols to provide multi-bit
masks.  Add separate-value and self-mutating methods to reverse the
order of bits within an integer.