Add methods for the head of a packed-Boolean collection
Add methods to inspect or remove the head of a packed-Boolean
collection, where the methods take the number of elements targeted as a
parameter.  Add a method to insert copies of another instance’s
elements at the head of a collection.  Add properties to measure the
remnant end-word and the full words.

Extend an integer protocol to assign to selected bits of an integer.
Extend another integer protocol to use an integer like a deque; pushing
new bits from either the most- or least-significant end to change the
value and returning the bits ejected from the other end (which can be
used to cascade a push to adjacent integers).