Add text representation for debugging
Add property to conform to the Standard Library manner to display
debugging output.  The format prints out all the bits of each
completely filled word, in hexadecimal, followed by the composite value
and count of trailing elements that don’t complete the last word (if
necessary).

Extend the standard string type with a method to create a copy of a
receiving string that is padded to a given character count, using
copies a given character to fill in from the left-side if needed.

Extend an integer protocol with a type-level property for the maximum
hexadecimal digit string count.  Extend it again with a instance-level
method to generate a maximum-length hexadecimal string representation
of the receiver.