Add basic Collection support to the packed-Boolean type
Add members to the packed-Boolean collection type to actually conform
with the most basic collection protocol.  A custom index type, with an
inner word index and a bit mask, is used instead of an Int offset.

Extend an integer protocol with a type-level property for the mask
value with only the most-significant bit set.