Optimize members of Collection for packed-Boolean storage
Override some properties of Collection to work on the internal
properties directly instead of keeping the default implementations,
which use (relatively) more expensive index creation and measuring
operations.  Override an initializer and a method of
RangeReplaceableCollection to set up elements in bulk, instead of the
one-element-at-a-time updates in their default implementations.

Within the corresponding tests, restore the tests removed in the
previous commit.